tests/wpt/tests/speculation-rules/prerender/restriction-service-worker-unregister.https.html

<!DOCTYPE html>
<title>ServiceWorkerRegistration.unregister in a prerendered page</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="/service-workers/service-worker/resources/test-helpers.sub.js"></script>
<script src="../resources/utils.js"></script>
<script src="resources/utils.js"></script>

<body>
<script>
setup(() => assertSpeculationRulesIsSupported());

const uid = token();

const PAGE_URL = `resources/service-worker-unregister.html?uid=${uid}`;
const WORKER_URL = 'resources/do-nothing-worker.js';

promise_test(async t => {
  const registration =
      await navigator.serviceWorker.register(WORKER_URL, {scope: PAGE_URL});
  t.add_cleanup(_ => registration.unregister());

  const bc = new PrerenderChannel('test-channel', uid);
  t.add_cleanup(_ => bc.close());

  const gotMessage = new Promise(resolve => {
    bc.addEventListener('message', e => {
      resolve(e.data);
    }, {
      once: true
    });
  });
  window.open(PAGE_URL, '_blank', 'noopener');

  const result = await gotMessage;
  const expected = [
    {event: 'started waiting ServiceWorkerRegistration.unregister',  prerendering: true},
    {event: 'prerendering change',                                   prerendering: false},
    {event: 'service worker unregistered',                           prerendering: false},
    {event: 'finished waiting ServiceWorkerRegistration.unregister', prerendering: false},
  ];
  assert_equals(result.length, expected.length);
  for (let i = 0; i < result.length; i++) {
    assert_equals(result[i].event, expected[i].event, `event[${i}]`);
    assert_equals(result[i].prerendering, expected[i].prerendering,
      `prerendering[${i}]`);
  }

  // Send a close signal to PrerenderEventCollector on the prerendered page.
  new PrerenderChannel('close', uid).postMessage('');
}, 'ServiceWorkerRegistration.unregister() should be deferred in a ' +
   'prerendered page');
</script>
